refactored grenade mocks in tg-frontend in separate file
